# By Cherisha Choukse
# Via Cherisha Choukse
* 'master' of https://github.com/mayurj07/TagIt:
  sql alt Table -comment
  comment functionalities- update comment , get all comments, delete comment

+++ src/main/java/edu/sjsu/cmpe275/tagit/models/Comment/Comment.java
@@ -18,7 +18,7 @@ public class Comment {
     @JsonProperty
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int  commentid;
+    private long  commentid;
 
     @JsonProperty
     @Column(name="Description")
@@ -42,7 +42,7 @@ public class Comment {
     }
 
 
-    public int getCommentid() {
+    public long getCommentid() {
         return commentid;
     }
